requirements.txt
# testing
pexpect-u
tox
pytest

# documentation
sphinx

# scripts
requests
